Convert PoC main into automated spec
